lustre: llite: Make iotrace logging quieter
Most of the time, we don't read any pages with readahead,
since we're moving through the window and aren't ready to
read more yet.  That's important for readahead debug, but
there's no need to log it for iotrace.  (This matters
because without this change, this message is the large
majority of iotrace messages.)

WC-bug-id: https://jira.whamcloud.com/browse/LU-15317
Lustre-commit: a91b5d4a990c6a870 ("LU-15317 llite: Make iotrace logging quieter")
Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/45887
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>